// types/enums.ts
export enum ServiceSpecType {
  TEXT = 'text',
  SELECT = 'select',
  MULTISELECT = 'multiselect',
  NUMBER = 'number',
  IMAGE = 'image',
}

export enum BOOKING_STATUS {
  PENDING = 'PENDING', // booking created, finding technician
  ACCEPTED = 'ACCEPTED', // technician accepted
  ENROUTE = 'ENROUTE', // technician on the way
  ARRIVED = 'ARRIVED', // technician reached customer location
  ONGOING = 'ONGOING', // work started
  COMPLETED = 'COMPLETED', // work done
  CANCELLED = 'CANCELLED', // cancelled by technician
}

export enum BOOKING_PAYMENT_STATUS {
  PENDING = 'PENDING',
  PAID = 'PAID',
  FAILED = 'FAILED',
}

export enum COMPLAINT_STATUS {
  PENDING = 'PENDING',
  RESOLVED = 'RESOLVED',
}

export enum PAYMENT_STATUS {
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
}

export enum PLATFORM_SETTING_KEYS {
  PLATFORM_FEE = 'PLATFORM_FEE',
  GST = 'GST',
  CURRENCY = 'CURRENCY',
  MIN_BOOKING_AMOUNT = 'MIN_BOOKING_AMOUNT',
  MAX_BOOKING_DISTANCE_KM = 'MAX_BOOKING_DISTANCE_KM',
  CUSTOMER_CANCELLATION_FEE = 'CUSTOMER_CANCELLATION_FEE',
  TECHNICIAN_COMMISSION = 'TECHNICIAN_COMMISSION',
  EMERGENCY_BOOKING_CHARGE = 'EMERGENCY_BOOKING_CHARGE',
  SUPPORT_PHONE = 'SUPPORT_PHONE',
  SUPPORT_EMAIL = 'SUPPORT_EMAIL',
  CUSTOMER_TOKEN_AMOUNT = 'CUSTOMER_TOKEN_AMOUNT',
}

export enum TECHNICIAN_LOCATION_TYPE {
  WORK = 'WORK',
}

export enum PAYOUT_ACCOUNT_TYPE {
  VPA = 'VPA',
  BANK_ACCOUNT = 'BANK_ACCOUNT',
}

export enum PAYOUT_ONBOARDING_STATUS {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  FAILED = 'FAILED',
}

export enum TECHNICIAN_PAYOUT_STATUS {
  PENDING = 'PENDING',
  PROCESSING = 'PROCESSING',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
  REVERSED = 'REVERSED',
}

export enum TECHNICIAN_STATUS {
  PENDING_APPROVAL = 'PENDING_APPROVAL',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
}

export enum JOB_STATUS {
  OFFLINE = 'OFFLINE',
  BUSY = 'BUSY',
  AVAILABLE = 'AVAILABLE',
}

export enum USER_ROLE {
  CUSTOMER = 'CUSTOMER',
  TECHNICIAN = 'TECHNICIAN',
  ADMIN = 'ADMIN',
}

export enum USER_STATUS {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  // REJECTED = 'REJECTED',
  BLOCKED = 'BLOCKED',
  PENDING = 'PENDING',
}